Reworking the backend into a single database object
